KYLIN: net: mucse: fix compilation errors
Mainline: KYLIN-only
Severity: Low

Task: #587923

The compilation error on i386：

CC net/ipv4/inet_timewait_sock.o
In file included from ../drivers/net/ethernet/mucse/rnpm/rnpm_common.c:15:
../drivers/net/ethernet/mucse/rnpm/rnpm_mbx_fw.h: In function ‘mbx_fw_req_set_reply’:
../drivers/net/ethernet/mucse/rnpm/rnpm_mbx_fw.h:868:32: error: right shift count >= width of type [-Werror=shift-count-overflow]
868 | req->reply_hi = (reply >> 32);
| ^~
cc1: all warnings being treated as errors
make[8]: *** [../scripts/Makefile.build:243: drivers/net/ethernet/mucse/rnpm/rnpm_common.o] Error 1
make[8]: *** Waiting for unfinished jobs....
CC [M] drivers/net/ethernet/grtnic_xgb/grtnic_macphy.o
CC drivers/hid/hid-samsung.o
CC [M] drivers/net/ethernet/mucse/rnp/rnp_lib.o
In file included from ../drivers/net/ethernet/mucse/rnpm/rnpm_debugfs.c:7:
../drivers/net/ethernet/mucse/rnpm/rnpm_mbx_fw.h: In function ‘mbx_fw_req_set_reply’:
../drivers/net/ethernet/mucse/rnpm/rnpm_mbx_fw.h:868:32: error: right shift count >= width of type [-Werror=shift-count-overflow]
868 | req->reply_hi = (reply >> 32);
| ^~
cc1: all warnings being treated as errors
In file included from ../drivers/net/ethernet/mucse/rnp/rnp_ethtool.c:18:
../drivers/net/ethernet/mucse/rnp/rnp_mbx_fw.h: In function ‘build_ddr_csl’:
../drivers/net/ethernet/mucse/rnp/rnp_mbx_fw.h:1155:52: error: right shift count >= width of type [-Werror=shift-count-overflow]
1155 | req->ddr_csl.ddr_phy_hi = (dma_phy >> 32);
| ^~
cc1: all warnings being treated as errors
make[8]: *** [../scripts/Makefile.build:243: drivers/net/ethernet/mucse/rnp/rnp_ethtool.o] Error 1
make[8]: *** Waiting for unfinished jobs....
make[8]: *** [../scripts/Makefile.build:243: drivers/net/ethernet/mucse/rnpm/rnpm_debugfs.o] Error 1
CC drivers/gpu/drm/drm_plane.o
make[7]: *** [../scripts/Makefile.build:480: drivers/net/ethernet/mucse/rnpm] Error 2

K2CI-Arch: All
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/170908
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>

+++ drivers/net/ethernet/mucse/rnpm/rnpm_mbx_fw.h
@@ -865,7 +865,7 @@ static inline void build_mbx_wol_set(struct mbx_fw_cmd_req *req,
 static inline void mbx_fw_req_set_reply(struct mbx_fw_cmd_req *req,
 					dma_addr_t reply)
 {
-	req->reply_hi = (reply >> 32);
+	req->reply_hi = (u32)((u64)(reply >> 32));
 	req->reply_lo = (reply) & 0xffffffff;
 }
